pyproject.toml
[project]
name = "outstanding-ai-engine"
version = "0.1.0"
description = "AI Engine for Outstanding AI debt collection platform"
requires-python = ">=3.12"

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # LLM Providers
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",        # OpenAI provider (fallback)
    "google-genai>=1.0,<2",
    "google-auth>=2.30,<3",

    # Security & Rate Limiting
    "slowapi>=0.1.9",  # Rate limiting for FastAPI

    # Utilities
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "nest-asyncio>=1.5.0",

    # Google API (for Vertex retry handling)
    "google-api-core>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.15",
    "pre-commit>=4.5.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
